Update CHANGELOG.md
All notable changes to the EntroPy Password Generator project are documented in this file. This project adheres to the Keep a Changelog standard, which ensures a structured and human-readable format for tracking changes. By following this approach, we provide clear visibility into the project's evolution, making it easier for users and contributors to understand what has been added, changed, or fixed in each release.