Update README.markdown
EntroPy Password Generator is a secure and customizable password generator developed in Python. It generates strong passwords with configurable character sets and calculates their entropy to evaluate robustness. The project provides 20 modes for secure password generation, ranging from 15 to 128 characters, with entropies between 95.10 bits and 811.50 bits. This exceeds the Proton© recommended minimum of 75 bits and the cryptographic strength best practices outlined in NIST SP 800-63B and NIST SP 800-132.

With a command-line interface, it is an ideal solution for developers, security enthusiasts, and anyone requiring robust passwords.